COMP: Explicitely onvert int to double during call of sin to remove ambiguity error.

+++ Base/Segmentation/VTree/Testing/itkSplineApprox1DTest.cxx
@@ -46,7 +46,7 @@ class MySA1DFunc:
       };
     const double & value( const int & x )
       {
-      cVal = vcl_sin(x);
+      cVal = vcl_sin((double)x);
       return cVal;
       };
   };
